TkinterFrontend/requirements.txt
altgraph==0.17.4
certifi==2024.8.30
charset-normalizer==3.4.0
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.6.0
urllib3==2.2.3
